These are the new versions Main and ExtractData
There are probably some changes to main that you made that I did not, such as changing the name of csv file to include both of our usernames. You will probably need to review that before submitting. But I think I got it working. It now outputs 1067 lines the way he wants it to